Refactor project hierarchy
