--- src/CinemaKeeper.Service/Adapters/IAdapter.cs
@@ -1,7 +1,0 @@
-﻿namespace CinemaKeeper.Service.Adapters
-{
-    internal interface IAdapter<out T>
-    {
-        T Convert();
-    }
-}
