--- src/CinemaKeeper.Service/Exceptions/InvalidUserLimitException.cs
@@ -1,6 +1,0 @@
-﻿using System;
-
-namespace CinemaKeeper.Service.Exceptions
-{
-    public class InvalidUserLimitException : Exception { }
-}
