--- src/CinemaKeeper.Service/Exceptions/InvalidVoiceChannelIdentifierException.cs
@@ -1,6 +1,0 @@
-﻿using System;
-
-namespace CinemaKeeper.Service.Exceptions
-{
-    public class InvalidVoiceChannelIdentifierException : Exception { }
-}
