--- src/CinemaKeeper.Service/Exceptions/VoiceChannelNotFoundException.cs
@@ -1,6 +1,0 @@
-﻿using System;
-
-namespace CinemaKeeper.Service.Exceptions
-{
-    public class VoiceChannelNotFoundException : Exception { }
-}
