--- src/CinemaKeeper.Service/GlobalSuppressions.cs
@@ -1,4 +1,0 @@
-﻿using System.Diagnostics.CodeAnalysis;
-
-[assembly: SuppressMessage("Performance", "CA1812",
-    Scope = "namespaceanddescendants", Target = "~N:CinemaKeeper.Service.Services")]
