--- src/CinemaKeeper.Service/Helpers/VoiceChannelIdType.cs
@@ -1,9 +1,0 @@
-﻿namespace CinemaKeeper.Service.Helpers
-{
-    internal enum VoiceChannelIdType
-    {
-        None,
-        Wildcard,
-        Id
-    }
-}
