+++ src/CinemaKeeper/Commands/Helpers/VoiceChannelIdType.cs
@@ -1,0 +1,8 @@
+﻿namespace CinemaKeeper.Commands.Helpers;
+
+internal enum VoiceChannelIdType
+{
+    None,
+    Wildcard,
+    Id
+}
