+++ src/CinemaKeeper/Exceptions/InvalidUserLimitException.cs
@@ -1,0 +1,5 @@
+﻿using System;
+
+namespace CinemaKeeper.Exceptions;
+
+public class InvalidUserLimitException : Exception { }
