+++ src/CinemaKeeper/Exceptions/InvalidVoiceChannelIdentifierException.cs
@@ -1,0 +1,5 @@
+﻿using System;
+
+namespace CinemaKeeper.Exceptions;
+
+public class InvalidVoiceChannelIdentifierException : Exception { }
