+++ src/CinemaKeeper/Exceptions/VoiceChannelNotFoundException.cs
@@ -1,0 +1,5 @@
+﻿using System;
+
+namespace CinemaKeeper.Exceptions;
+
+public class VoiceChannelNotFoundException : Exception { }
